+++ integrations/selenium_grid/ReadMe.md
@@ -1,29 +1,1 @@
-## The Selenium Grid Hub
-
-The Selenium Grid Hub lets you distribute tests to run in parallel across multiple node machines. Each node machine can then run its own allocation of tests. This allows you to run a large suite of tests very quickly.
-
-### Running the Selenium Grid Hub
-
-The following commands will work once you've installed seleniumbase, which comes with the seleniumbase console scripts interface.
-
-Grid Hub server controls:
-```bash
-seleniumbase grid-hub {start|stop|restart} [OPTIONS]
-```
-Options:
-* ``-v``, ``--verbose``  (Increases verbosity of logging output.)
-
-Grid node server controlls:
-```bash
-seleniumbase grid-node {start|stop|restart} --hub=[HUB_IP] [OPTIONS]
-```
-Options:
-* ``-v``, ``--verbose``  (Increases verbosity of logging output.)
-* ``--hub=[HUB_IP]`` (Specifies the Grid Hub to connect to. Default: "127.0.0.1".)
-
-When the Grid Hub Console is up and running, you'll be able to find it here: [http://127.0.0.1:4444/grid/console](http://127.0.0.1:4444/grid/console)
-
-
-#### More info about the Selenium Grid Hub can be found here:
-* [https://github.com/SeleniumHQ/selenium/wiki/Grid2](https://github.com/SeleniumHQ/selenium/wiki/Grid2)
-* [https://github.com/SeleniumHQ/selenium/wiki](https://github.com/SeleniumHQ/selenium/wiki/Grid2)
+### The ReadMe for the Selenium Grid Hub Launcher has been moved to: [seleniumbase/utilities/selenium_grid/ReadMe.md](https://github.com/seleniumbase/SeleniumBase/blob/master/seleniumbase/utilities/selenium_grid/ReadMe.md) and all related code has been moved to [seleniumbase/utilities/selenium_grid](https://github.com/seleniumbase/SeleniumBase/blob/master/seleniumbase/utilities/selenium_grid)
