+++ seleniumbase/utilities/selenium_grid/font_color
